# Speed up `Transaction Info` Cucumber test