--- a/comms/dht/src/proto/mod.rs
+++ b/comms/dht/src/proto/mod.rs
@@ -23,6 +23,7 @@
 use crate::proto::{dht::JoinMessage, envelope::Network};
 use rand::{rngs::OsRng, RngCore};
 use std::{convert::TryInto, fmt};
+use tari_common::Network as GlobalNetwork;
 use tari_comms::{
     multiaddr::Multiaddr,
     peer_manager::{NodeId, Peer, PeerFeatures, PeerFlags},
@@ -64,6 +65,18 @@ impl envelope::Network {
 
     pub fn is_localtest(self) -> bool {
         matches!(self, Network::LocalTest)
+    }
+}
+
+impl From<GlobalNetwork> for Network {
+    fn from(gn: GlobalNetwork) -> Self {
+        match gn {
+            GlobalNetwork::MainNet => Network::MainNet,
+            GlobalNetwork::Rincewind => Network::TestNet,
+            GlobalNetwork::LocalNet => Network::LocalTest,
+            GlobalNetwork::Ridcully => Network::Ridcully,
+            GlobalNetwork::Stibbons => Network::Stibbons,
+        }
     }
 }