## Description
The `Transaction Info` was very slow take 12 minutes to complete on average.

This PR fixes a number of issues which brings the running time down to 40 seconds

1. Reduce the `broadcast_monitoring_timeout` from 60 seconds to 3 seconds to make the broadcast protocols more responsiv
2. Fix the `Network` enum used by the common configuration and the DhtConfig object to properly reflect that the cucumber test is using a LocalTest network where you can discover local IPv4 addresses
3. Reduce the Liveness service `auto_ping_interval` to 15 seconds to reduce the time it takes for the second base node to detect the new blocks that were mined on the seed node.

The timing was tested using the udated version of this cucumber test found in https://github.com/tari-project/tari/pull/2640

## How Has This Been Tested?
This is a test

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
* [x] Bug fix (non-breaking change which fixes an issue)
* [ ] New feature (non-breaking change which adds functionality)
* [ ] Breaking change (fix or feature that would cause existing functionality to change)
* [ ] Feature refactor (No new feature or functional changes, but performance or technical debt improvements)
* [ ] New Tests
* [ ] Documentation

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
* [x] I'm merging against the `development` branch.
* [x] I ran `cargo-fmt --all` before pushing.
* [x] I ran `cargo test` successfully before submitting my PR.
* [x] I have squashed my commits into a single commit.
* [ ] My change requires a change to the documentation.
* [ ] I have updated the documentation accordingly.
* [ ] I have added tests to cover my changes.